Update jelastic.json
Actually Nginx is used as a php and ruby app server + balancing web-server;
Apache - as app server for php, ruby and python;
since it is allowed to specify just a one runtime/role for each of them - I've stated the PHP one for Apache and Ruby for NGINX server. Nevertheless, it woulf be nice if you could find a way to extend their specification or allow to make dublicates.
Please reply if you see any more mistakes.